Fixed up output, added section for freshclam (still todo)

+++ configure-clamd.sh
@@ -19,15 +19,17 @@
 #This script is for installing and configuring clam-server (clamd) on Fedora
 
 #Make this more pretty by adding an extra blank line at the beginning
-clear
 echo ""
 
 #Print help, if requested
 if [ "$1" == "help" -o "$1" == "-help" -o "$1" == "--help" -o "$1" == "-h" -o "$1" == "--h" ]
 then
-	echo "This script installs and configures clamav-server (clamvd) on Fedora to listen on a local port."
-	echo "Pass in the user to run clamd as and the port to run on, else these will default to 'clamav' on port '3310'. I.e.:"
-	echo "su -c ./configure-clamd.sh [username] [port]"
+	echo "This script installs and configures clamav-server (clamd) on Fedora."
+	echo ""
+	echo "You may pass the user to run clamd, as well as the port to run on:"
+	echo ' su -c "./configure-clamd.sh [username] [port]"'
+	echo ""
+	echo "Else these will default to user 'clamav' on port '3310'."
 	echo ""
 	exit 0
 fi
@@ -72,28 +74,18 @@ then
 	CLAMD_PORT="$2"
 fi
 
-#Check to see if the port is already in use, if so, increment by one until we find something that's free
-while [ -n "`netstat -ltn |grep ":$CLAMD_PORT"`" ]
-do
-if [ -n "`netstat -ltn |grep ":$CLAMD_PORT"`" ]
-then
-	CLAMD_PORT=$(($CLAMD_PORT+1))
-fi
-done
-
 #Checking to see if required packages are installed or not
 echo "Checking for required packages.."
-echo ""
 if [ -n "`rpm -qa |grep clamav`" -a "`rpm -qa |grep clamav-update`" -a "`rpm -qa |grep clamav-server`" ]
 then
-	echo "Required packages already installed, continuing.."
+	echo "Required packages already installed."
 	echo ""
 else
 
 	#Install required packages
 	echo "Installing required clamav packages.."
 	echo ""
-	yum -y install clamav clamav-server clamav-update
+	yum -yq install clamav clamav-server clamav-update
 	echo ""
 
 	#Check that the install was successful (or already installed)
@@ -115,8 +107,7 @@ CLAMD_VERSION="`rpm -qa |grep clamav-server |awk -F "-" {'print $3'} 2>/dev/null
 
 #Create clamav user if doesn't exist
 #This should be the user who wants to talk to clamd, else user clamav must have read (and possibly write) access on the files.
-echo "Checking for clamav user, $CLAMD_USER."
-echo ""
+echo "Checking for clamav user, $CLAMD_USER.."
 
 if [ -z "`id $CLAMD_USER 2>/dev/null`" ]
 then
@@ -129,11 +120,11 @@ then
 		echo ""
 		exit 1
 	else
-		echo "Created new user, $CLAMD_USER."
+		echo "Created new user."
 		echo ""
 	fi
 else
-	echo "User already exists, continuing."
+	echo "User already exists."
 	echo ""
 fi
 
@@ -142,10 +133,24 @@ CLAMD_CONFIG="/etc/clamd.d/$CLAMD_USER.conf"
 
 #Copy and configure clamd configuration file
 echo "Configuring clamd to do all the right things.."
-echo ""
 
-#Try to remove existing config, whether it exists or not because 'cp' is aliased with -i
-rm -f $CLAMD_CONFIG 2>/dev/null
+if [ -e $CLAMD_CONFIG ]
+then
+	echo "Instance of clamd already exists, clobbering.."
+	#Stop existing daemon to enable new one
+	/etc/init.d/clamd.$CLAMD_USER stop &>/dev/null
+	#Remove existing config because 'cp' is aliased with -i and we don't want a prompt
+	rm -f $CLAMD_CONFIG 2>/dev/null
+fi
+
+#Check to see if the port is already in use, if so, increment by one until we find something that's free
+while [ -n "`netstat -ltn |grep ":$CLAMD_PORT"`" ]
+do
+if [ -n "`netstat -ltn |grep ":$CLAMD_PORT"`" ]
+then
+	CLAMD_PORT=$(($CLAMD_PORT+1))
+fi
+done
 
 #Check that the template clamd.conf exists
 if [ -f /usr/share/doc/clamav-server-$CLAMD_VERSION/clamd.conf ]
@@ -159,11 +164,9 @@ then
 	sed -i 's/^#LogFile/LogFile/' $CLAMD_CONFIG
 	sed -i 's/^#PidFile/PidFile/' $CLAMD_CONFIG
 	sed -i 's/^LocalSocket/#LocalSocket/' $CLAMD_CONFIG
-	sed -i 's/^#TCPSocket/TCPSocket/' $CLAMD_CONFIG
+	sed -i 's/^#TCPSocket\ 3310/TCPSocket\ '$CLAMD_PORT'/' $CLAMD_CONFIG
 	sed -i 's/^#TCPAddr/TCPAddr/' $CLAMD_CONFIG
 	sed -i 's/<USER>/'$CLAMD_USER'/' $CLAMD_CONFIG
-	echo "Complete."
-	echo ""
 else
 	echo "Could not find clamd.conf template under /usr/share/doc/clamav-server-$CLAMD_VERSION/, sorry."
 	echo ""
@@ -171,12 +174,13 @@ else
 	echo ""
 	exit 1
 fi
+echo "Done."
+echo ""
 
 #Copy and configure clamd for log rotation
 if [ -d /etc/logrotate.d ]
 then
 	echo "Configuring log rotation for clamd.."
-	echo ""
 	CLAMD_LOGROTATE=/etc/logrotate.d/clamd-$CLAMD_USER
 
 	if [ -f /usr/share/doc/clamav-server-$CLAMD_VERSION/clamd.logrotate ]
@@ -185,8 +189,6 @@ then
 		rm -f $CLAMD_LOGROTATE 2>/dev/null
 		cp -f /usr/share/doc/clamav-server-$CLAMD_VERSION/clamd.logrotate $CLAMD_LOGROTATE
 		sed -i 's/clamd.<SERVICE>/clamd.'$CLAMD_USER'/' $CLAMD_LOGROTATE
-		echo "Complete."
-		echo ""
 	else
 		echo "Could not find logrotate template under /usr/share/doc/clamav-server-$CLAMAV_VERSION, sorry."
 		echo ""
@@ -194,13 +196,14 @@ then
 		echo ""
 	fi
 fi
+echo "Done."
+echo ""
 
 #Set variable for clamd sysconfig
 CLAMD_SYSCONFIG="/etc/sysconfig/clamd.$CLAMD_USER"
 
 #Configuring clamd under sysconfig
 echo "Configuring clamd under syconfig.."
-echo ""
 
 #Try to remove existing config, whether it exists or not because 'cp' is aliased with -i
 rm -f $CLAMD_SYSCONFIG 2>/dev/null
@@ -212,8 +215,6 @@ then
         cp -f /usr/share/doc/clamav-server-$CLAMD_VERSION/clamd.sysconfig $CLAMD_SYSCONFIG 2>/dev/null
         sed -i 's/<SERVICE>/'$CLAMD_USER'/' $CLAMD_SYSCONFIG
         sed -i 's/^#CLAMD/'CLAMD'/' $CLAMD_SYSCONFIG
-        echo "Complete."
-        echo ""
 else
         echo "I can't find the sysconfig template under /usr/share/doc/clamav-server-$CLAMD_VERSION/, sorry."
         echo ""
@@ -221,13 +222,14 @@ else
         echo ""
         exit 1
 fi
+echo "Done."
+echo ""
 
 #Set variable for clamd init script
 CLAMD_INIT="/etc/init.d/clamd.$CLAMD_USER"
 
 #Configuring clamd init script
 echo "Configuring clamd init script.."
-echo ""
 
 #Try to remove existing config, whether it exists or not because 'cp' is aliased with -i
 rm -f $CLAMD_INIT 2>/dev/null
@@ -238,7 +240,7 @@ then
         #Copy over the init script
         cp -f /usr/share/doc/clamav-server-$CLAMD_VERSION/clamd.init $CLAMD_INIT 2>/dev/null
         sed -i 's/<SERVICE>/'$CLAMD_USER'/' $CLAMD_INIT
-	ln -s /usr/sbin/clamd /usr/sbin/clamd.$CLAMD_USER
+	ln -s /usr/sbin/clamd /usr/sbin/clamd.$CLAMD_USER 2>/dev/null
 	/sbin/chkconfig clamd.$CLAMD_USER on
 	#Check that was successful
 	if [ $? != 0 ]
@@ -248,9 +250,6 @@ then
 		echo "Exiting."
 		echo ""
 		exit 1
-	else
-	        echo "Complete."
-	        echo ""
 	fi
 else
         echo "I can't find the sysconfig template under /usr/share/doc/clamav-server-$CLAMD_VERSION/, sorry."
@@ -259,8 +258,16 @@ else
         echo ""
         exit 1
 fi
+echo "Done."
+echo ""
 
+#Configure freshclam
+echo "Configuring freshclam, the clamav updater.."
+#TODO
+echo "Done."
+echo ""
 
+echo "Creating required directories and starting service.."
 #Setup logs
 CLAMD_LOG=/var/log/clamd.$CLAMD_USER
 touch $CLAMD_LOG
@@ -269,18 +276,20 @@ chmod 0620 $CLAMD_LOG
 
 #Setup run socket
 CLAMD_PID=/var/run/clamd.$CLAMD_USER
-mkdir $CLAMD_PID
+mkdir $CLAMD_PID 2>/dev/null
 chown $CLAMD_USER:$CLAMD_USER $CLAMD_PID/
 
 #Start services
+echo ""
 /etc/init.d/clamd.$CLAMD_USER start
 if [ $? != 0 ]
 then
 	echo "Could not start service, sorry."
-	echo ""
-	echo "Continuing"
+	echo "Continuing."
 	echo ""
 fi
+echo "Done."
+echo ""
 
 #Print summary
 echo "The clamd service has been successfully installed and configured with:"
